# lib/core/v1/swagger.yaml
swagger: '2.0'
info:
  description: >-
    The document shows Elle's endpoints
  version: 1.0.0
  title: Elle
  contact:
    email: valerio.cogni@n-3.it
schemes:
  - https
  - http
host: brokeragesystem.it
basePath: /core/v1
...
paths:
...
  /permission:
    get:
      tags:
        - CORE
      summary: Person's permits
      description: ""
      operationId: permission
      produces:
        - application/json
      parameters:
        - name: gui
          in: query
          description: Gui's Name
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: "#/definitions/ResponsePermission"
        "401":
          description: Unauthorized
...
definitions:
...
  ResponsePermission:
    type: object
    properties:
      result:
        type: boolean
      data:
        $ref: "#/definitions/GuiPermission"
...
  GuiPermission:
    type: object 
    properties:
      permission:
         $ref: "#/definitions/Permission"
...
  Permission:
    type: object
    properties:
      execute: 
        type: boolean
      read: 
        type: boolean
      write: 
        type: boolean
      active:
        type: boolean
...